Time: 7 ms (80.47%), Space: 5.9 MB (48.50%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -6,7 +6,7 @@ public:
         }
         
         int num1=x;
-        long long num2=0;
+        long num2=0;
         
         while(num1){
             num2 = num2*10 + num1%10 ;
